write more tests

+++ src/test/java/com/jianglibo/wx/katharsis/rest/group/TestGrouptApi.java
@@ -1,4 +1,4 @@
-package com.jianglibo.wx.katharsis.rest;
+package com.jianglibo.wx.katharsis.rest.group;
 
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.Assert.assertThat;
@@ -8,6 +8,7 @@ import java.io.IOException;
 
 import org.junit.Before;
 import org.junit.Test;
+import org.springframework.beans.factory.annotation.Autowired;
 
 import com.fasterxml.jackson.core.JsonParseException;
 import com.fasterxml.jackson.databind.JsonMappingException;
@@ -15,14 +16,18 @@ import com.jianglibo.wx.JsonApiPostBodyWrapper;
 import com.jianglibo.wx.JsonApiPostBodyWrapperBuilder;
 import com.jianglibo.wx.KatharsisBase;
 import com.jianglibo.wx.config.JsonApiResourceNames;
+import com.jianglibo.wx.domain.GroupUserRelation;
 import com.jianglibo.wx.katharsis.dto.GroupDto;
+import com.jianglibo.wx.repository.GroupUserRelationRepository;
 
 public class TestGrouptApi  extends KatharsisBase {
+	
+	@Autowired
+	private GroupUserRelationRepository gurRepo;
 	
 	@Before
 	public void b() throws JsonParseException, JsonMappingException, IOException {
 		initTestUser();
-		delteAllGroups();
 	}
 	
 	@Test
@@ -38,6 +43,7 @@ public class TestGrouptApi  extends KatharsisBase {
 		String bodys = indentOm.writeValueAsString(body);
 		writeDto(bodys, getResourceName(), "postcontent");
 		
+		// every one is capable to create any number of groups.
 		response = postItemWithContent(bodys, jwt1);
 		
 		writeDto(response, getResourceName(), ActionNames.POST_RESULT);
@@ -48,12 +54,22 @@ public class TestGrouptApi  extends KatharsisBase {
 		assertThat(newGroup.getThumbUrl(), equalTo("abc"));
 		assertTrue(newGroup.isOpenToAll());
 		
+		// only the administrator can get group list. 
 		response = requestForBody(jwt1, getBaseURI());
 		writeDto(response, getResourceName(), ActionNames.GET_LIST);
 		
 		response = requestForBody(jwt1, getItemUrl(newGroup.getId()));
 		
 		writeDto(response, getResourceName(), ActionNames.GET_ONE);
+		
+		// access denied.
+		response = deleteByExchange(jwt2, getItemUrl(newGroup.getId()));
+		assertAccessDenied(response);
+		
+		// add user to group.
+		GroupUserRelation gur = new GroupUserRelation(groupRepo.findOne(newGroup.getId()), user2);
+		gurRepo.save(gur);
+		
 		deleteByExchange(jwt1, getItemUrl(newGroup.getId()));
 	}
